requirements.txt
# Core packages
fastparquet==2023.10.0
numpy==1.23.5
pandas==1.5.3
pyarrow==11.0.0
python-dotenv==1.0.1
requests==2.31.0

# Google Cloud SDK
google-auth==2.29.0
google-cloud-bigquery==3.11.4
google-cloud-storage==2.16.0

# Apache Airflow
apache-airflow==2.8.1
psycopg2-binary

# Testing
pytest==8.2.1
responses==0.25.0  # For mocking requests in tests

# Linting & formatting
black==24.3.0
flake8==7.0.0
pylint==3.1.0